+++ NEWS.md
@@ -1,5 +1,7 @@
 # remotes (development version)
 
+* Replaced `mockery::stub()` with `testthat::local_mocked_bindings()`. Removed {mockery} as a suggested package.
+
 # remotes 2.5.0
 
 * `install_github()` now uses credentials from the git